+++ src/app/App.js
@@ -2,8 +2,8 @@ import React from "react";
 import { Route, Switch } from "react-router-dom";
 import Login from "./layouts/login";
 import Main from "./layouts/main";
-import NavBar from "./components/navBar";
-import UsersList from "./components/usersList";
+import NavBar from "./components/ui/navBar";
+import Users from "./layouts/users";
 
 function App() {
   return (
@@ -11,7 +11,7 @@ function App() {
       <NavBar />
       <Switch>
         <Route path="/" exact component={Main} />
-        <Route path="/users/:userId?" component={UsersList} />
+        <Route path="/users/:userId?" component={Users} />
         <Route path="/login" component={Login} />
       </Switch>
     </div>
