+++ src/app/components/common/bookmark.jsx
@@ -8,8 +8,8 @@ const BookMark = ({ status, ...rest }) => {
     </button>
   );
 };
-
 BookMark.propTypes = {
   status: PropTypes.bool,
 };
+
 export default BookMark;
