+++ src/app/components/common/groupList.jsx
@@ -8,24 +8,7 @@ const GroupList = ({
   onItemSelect,
   selectedItem,
 }) => {
-  if (Array.isArray(items)) {
-    return (
-      <ul className="list-group">
-        {items.map((item) => (
-          <li
-            key={item[valueProperty]}
-            className={
-              "list-group-item" + (item === selectedItem ? " active" : "")
-            }
-            onClick={() => onItemSelect(item)}
-            role="button"
-          >
-            {item[contentProperty]}
-          </li>
-        ))}
-      </ul>
-    );
-  } else {
+  if (!Array.isArray(items)) {
     return (
       <ul className="list-group">
         {Object.keys(items).map((item) => (
@@ -44,13 +27,27 @@ const GroupList = ({
       </ul>
     );
   }
+  return (
+    <ul className="list-group">
+      {items.map((item) => (
+        <li
+          key={item[valueProperty]}
+          className={
+            "list-group-item" + (item === selectedItem ? " active" : "")
+          }
+          onClick={() => onItemSelect(item)}
+          role="button"
+        >
+          {item[contentProperty]}
+        </li>
+      ))}
+    </ul>
+  );
 };
-
 GroupList.defaultProps = {
   valueProperty: "_id",
   contentProperty: "name",
 };
-
 GroupList.propTypes = {
   items: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
   valueProperty: PropTypes.string.isRequired,
