+++ src/app/components/common/table/table.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
-import TableHeader from "./tableHeader";
 import TableBody from "./tableBody";
+import TableHeader from "./tableHeader";
 
 const Table = ({ onSort, selectedSort, columns, data, children }) => {
   return (
@@ -15,7 +15,6 @@ const Table = ({ onSort, selectedSort, columns, data, children }) => {
     </table>
   );
 };
-
 Table.propTypes = {
   onSort: PropTypes.func,
   selectedSort: PropTypes.object,
