+++ src/app/components/common/table/tableBody.jsx
@@ -6,7 +6,6 @@ const TableBody = ({ data, columns }) => {
   const renderContent = (item, column) => {
     if (columns[column].component) {
       const component = columns[column].component;
-
       if (typeof component === "function") {
         return component(item);
       }
