+++ src/app/components/common/table/tableHeader.jsx
@@ -22,6 +22,7 @@ const TableHeader = ({ onSort, selectedSort, columns }) => {
     }
     return null;
   };
+
   return (
     <thead>
       <tr>
@@ -44,7 +45,6 @@ const TableHeader = ({ onSort, selectedSort, columns }) => {
     </thead>
   );
 };
-
 TableHeader.propTypes = {
   onSort: PropTypes.func.isRequired,
   selectedSort: PropTypes.object.isRequired,
