+++ src/app/components/page/userPage/index.js
@@ -1,0 +1,3 @@
+import UserPage from "./userPage";
+
+export default UserPage;
