+++ src/app/components/page/usersListPage/index.js
@@ -1,0 +1,3 @@
+import UsersListPage from "./usersListPage";
+
+export default UsersListPage;
