+++ src/app/components/ui/quality/qualitiesList.jsx
@@ -1,19 +1,19 @@
 import React from "react";
-import Quality from "./qualitie";
 import PropTypes from "prop-types";
+import Quality from "./quality";
 
 const QualitiesList = ({ qualities }) => {
   return (
     <>
       {qualities.map((qual) => (
-        <Quality {...qual} key={qual._id} />
+        <Quality key={qual._id} {...qual} />
       ))}
     </>
   );
 };
 
 QualitiesList.propTypes = {
-  qualities: PropTypes.array.isRequired,
+  qualities: PropTypes.array,
 };
 
 export default QualitiesList;
