+++ src/app/components/ui/quality/quality.jsx
@@ -1,13 +1,13 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-const Qualities = ({ color, name }) => {
+const Quality = ({ color, name }) => {
   return <span className={"badge m-1 bg-" + color}>{name}</span>;
 };
-
-Qualities.propTypes = {
+Quality.propTypes = {
   color: PropTypes.string.isRequired,
   name: PropTypes.string.isRequired,
+  _id: PropTypes.string.isRequired,
 };
 
-export default Qualities;
+export default Quality;
