+++ src/app/layouts/users.jsx
@@ -1,137 +1,12 @@
-import React, { useState, useEffect } from "react";
-import PropTypes from "prop-types";
-import { paginate } from "../utils/paginate";
-import Pagination from "../components/pagination";
-import GroupList from "../components/groupList";
-import api from "../api";
-import SearchStatus from "../components/searchStatus";
-import UserTable from "../components/usersTable";
-import _ from "lodash";
+import React from "react";
+import { useParams } from "react-router-dom";
+import UserPage from "../components/page/userPage/userPage";
+import UsersListPage from "../components/page/usersListPage/usersListPage";
 
 const Users = () => {
-  const [currentPage, setCurrentPage] = useState(1);
-  const [searchUser, setSearchUser] = useState();
-  const [professions, setProfessions] = useState();
-  const [selectedProf, setSelectedProf] = useState();
-  const [sortBy, setSortBy] = useState({ iter: "name", order: "asc" });
-
-  const [users, setUsers] = useState();
-  useEffect(() => {
-    api.users.fetchAll().then((data) => setUsers(data));
-  }, []);
-  const handleDelete = (userId) => {
-    setUsers(users.filter((user) => user._id !== userId));
-  };
-  const handleToggleBookMark = (id) => {
-    setUsers(
-      users.map((user) => {
-        if (user._id === id) {
-          return { ...user, bookmark: !user.bookmark };
-        }
-        return user;
-      })
-    );
-    console.log(id);
-  };
-
-  const pageSize = 8;
-  useEffect(() => {
-    api.professions.fetchAll().then((data) => setProfessions(data));
-  }, []);
-  useEffect(() => {
-    setCurrentPage(1);
-    setSearchUser("");
-  }, [selectedProf]);
-
-  const handleProfessionSelect = (item) => {
-    setSelectedProf(item);
-  };
-
-  const handleChangeSearch = ({ target }) => {
-    setSearchUser(target.value);
-  };
-
-  const handlePageChange = (pageIndex) => {
-    setCurrentPage(pageIndex);
-  };
-
-  const handleSort = (item) => {
-    setSortBy(item);
-  };
-
-  if (users) {
-    let filteredUsers;
-    if (searchUser) {
-      filteredUsers = users.filter((user) => {
-        const regExp = new RegExp(`${searchUser}`);
-        return regExp.test(user.name);
-      });
-    } else {
-      filteredUsers = selectedProf
-        ? users.filter(
-            (user) =>
-              JSON.stringify(user.profession) === JSON.stringify(selectedProf)
-          )
-        : users;
-    }
-    const count = filteredUsers.length;
-    const sortedUsers = _.orderBy(filteredUsers, [sortBy.path], [sortBy.order]);
-    const usersCrop = paginate(sortedUsers, currentPage, pageSize);
-    const clearFilter = () => {
-      setSelectedProf();
-    };
-    return (
-      <div className="d-flex">
-        {professions && (
-          <div className="d-flex flex-column flex-shrink-0 p-3">
-            <GroupList
-              selectedItem={selectedProf}
-              items={professions}
-              onItemSelect={handleProfessionSelect}
-            />
-            <button className="btn btn-secondary mt-2" onClick={clearFilter}>
-              {" "}
-              Очистить
-            </button>
-          </div>
-        )}
-        <div className="d-flex flex-column">
-          <SearchStatus length={count} />
-          <div className="input-group mb-3">
-            <input
-              type="text"
-              id="site-search"
-              className="form-control"
-              value={searchUser}
-              onChange={handleChangeSearch}
-              onClick={clearFilter}
-            />
-          </div>
-          {count > 0 && (
-            <UserTable
-              users={usersCrop}
-              onSort={handleSort}
-              selectedSort={sortBy}
-              onDelete={handleDelete}
-              onToggleBookMark={handleToggleBookMark}
-            />
-          )}
-          <div className="d-flex justify-content-center">
-            <Pagination
-              itemsCount={count}
-              pageSize={pageSize}
-              currentPage={currentPage}
-              onPageChange={handlePageChange}
-            />
-          </div>
-        </div>
-      </div>
-    );
-  }
-  return "loading...";
-};
-Users.propTypes = {
-  users: PropTypes.array,
+  const params = useParams();
+  const { userId } = params;
+  return <>{userId ? <UserPage userId={userId} /> : <UsersListPage />}</>;
 };
 
 export default Users;
